Fixed occasional signal dropouts when using a Futaba T6EX.
Previously, spektrum receivers would briefly stop receiving signals
whenever the ppm value of any channel was increased quickly enough.
This was caused by the fact that the serial data packet was sent
as soon as seven channels were received from ppm.
However, since ppm signals have varying lengths, this caused
inconsistent timing of the serial output. And, as it turns out,
spektrum receivers expect a packet ~45 times per second.
The T6EX operates slighty slower, at 44.8Hz. When combined
with a sudden increase in packet length (e.g. when flipping a switch
that instantly moves a channel from min to max), this would result
in a particularly long pause between two serial packets.

The issue has been fixed by sending the serial output after a fixed
waiting period after the ppm frame has started. The delay can be
set in Defines.h.
Additionally, this patch has also reduced the overall latency:
The waiting perioud was chosen to send the serial packet after the
first four channels of the ppm signal have been received.
The values of the less important channels 5 and 6 are taken from
the last ppm frame. This behaviour trades 1-2 ms of latency for the
first four channels for about 22 ms of the last two. In my opinion,
a worthwhile trade. The behaviour can be changed in Defines.h by
adjusting the waiting period and the alarm threshold.

The alarm threshold is a safeguard: It lights up the indicator LED
when a serial packet is sent with less than a preset amount of channels
received from the current ppm frame. The default is set to 4.
This in-field checking if the waiting period is too small, which would cause
20 ms of latency on an important channel, such as channel 4.