Re-added blinking when binding to the LEDController.
The LEDController is no longer using a hardware timer.
This was done because the timer that was used, timer0, is also used
for millis() and micros(), which are used in the implementation
of the waiting period for sending serial output.
Additionally, the software-based implementation removes latency of the
error led.